Add loadTasks() function to TaskManager.java
Add a function fromFileFormat() in Task.java to support the loadTasks()
function. fromFileFormat takes in a line from the TaskList.txt file,
reads the line and returns the correct Event, Deadline or Todo object
depending on what the task is.

Update constructor of Task.java to take in a boolean isDone param
to support the fromFileFormat() function.

Update ToDo.java, Deadline.java and Event.java to have their respective
toFileFormat() function which converts the event into a file format to
be stored in TaskList.txt